Awesome Pull Request (#45)
* Update README

* Delete unneeded file

* Update language stuff

* Update disclaimer (1/2)

* Update disclaimer (2/2)

It still need to be updated on some languages, but that's all I could do for now :b

* Oops

I forgot to include this in the last commit :b

* Update README

I think you should include all the colaborators in the readme as well, not only Elijah

Btw, I'm arranging them in the order in which the languages ​​are shown (excluding Thomleg :b)

* Clearing all Thomleg traces...

* Remove Briccmii

C'mon, was it even necessary to include it in the first place? Hekate can also install AutoRCM by itself...

* Put payloads in individual files

* Add fusee-primary for Atmosphère 0.8.3

* Add proper contribuiting instructions

Plz, improve my grammar on the CONTRIBUITING file, english is hard >.<

* :facepalm: x2, maintaing

English is hard x2

* Bump version to 1.4

Because, why not? :b